refactor: Move links constants to index
