+++ apps/web/common/components/AdminLayout/Header.tsx
@@ -6,9 +6,10 @@ import { APP_NAME } from "@repo/constants"
 import {
   LINK_ACCOUNT,
   LINK_CREATE_ACCOUNT,
+  LINK_HOME,
   LINK_LOGIN,
   LINK_LOGOUT,
-} from "@/common/constants/links"
+} from "@/common/constants"
 import Link from "next/link"
 import { WidthWrapper } from "@/common/components/WidthWrapper"
 import { cn } from "@/common/helpers/cn"
@@ -50,7 +51,7 @@ const items = [
   },
   {
     name: "Guest Account",
-    href: "/",
+    href: LINK_HOME,
     selected: false,
   },
 ]
