+++ apps/web/common/components/HostLayout/NewListingRentalHeader.tsx
@@ -6,9 +6,10 @@ import { APP_NAME } from "@repo/constants"
 import {
   LINK_ACCOUNT,
   LINK_CREATE_ACCOUNT,
+  LINK_HOME,
   LINK_LOGIN,
   LINK_LOGOUT,
-} from "@/common/constants/links"
+} from "@/common/constants"
 import Link from "next/link"
 import { WidthWrapper } from "@/common/components/WidthWrapper"
 import { cn } from "@/common/helpers/cn"
@@ -49,7 +50,6 @@ function NewListingHeader({
 }) {
   const session = useSessionStore()
   const ASSET_ROOT = "/assets"
-  const current = "Hosting Account"
 
   const renderTransition = (children: React.ReactNode) => (
     <Transition
@@ -75,7 +75,7 @@ function NewListingHeader({
       <WidthWrapper width={contentWidth}>
         <nav className="flex items-center py-2 w-full">
           <div className="flex gap-2 items-center">
-            <Link href="/" className="">
+            <Link href={LINK_HOME} className="">
               <Image
                 className="h-10 w-auto"
                 src={Logo}
