+++ apps/web/common/constants/index.ts
@@ -1,6 +1,7 @@
 export * from "./page-titles"
 export * from "./api-routes"
 export * from "./time"
+export * from "./links"
 
 export const LOGO_SINGLE_IMAGE = "/logo-single.png"
 export const SUBMIT = "Submit"
