+++ apps/web/common/constants/links.ts
@@ -1,5 +1,5 @@
 export const LINK_CREATE_ACCOUNT = "/create-account"
-export const LINK_HOME = "/home"
+export const LINK_HOME = "/"
 export const LINK_LOGIN = "/login"
 export const LINK_SIGNUP = "/signup"
 export const LINK_LOGOUT = "/logout"
