+++ apps/web/module/AccountSettings/PersonalInfo/index.tsx
@@ -1,6 +1,6 @@
 "use client"
 import AccountSettingWrapper from "@/module/AccountSettings/components/AccountSettingWrapper"
-import { LINK_ACCOUNT } from "@/common/constants/links"
+import { LINK_ACCOUNT } from "@/common/constants"
 import React from "react"
 import LegalName from "./LegalName"
 import EmailAddress from "./EmailAddress"
