+++ apps/web/module/Authentication/Login/index.tsx
@@ -3,11 +3,12 @@ import React from "react"
 import LoginForm from "@/module/Authentication/components/LoginForm"
 import AuthContainer from "@/common/components/AuthContainer"
 import { useRouter } from "next/navigation"
+import { LINK_HOME } from "@/common/constants"
 const Login = () => {
   const router = useRouter()
 
   return (
-    <AuthContainer title="Login or sign up" onBack={() => router.push("/")}>
+    <AuthContainer title="Login or sign up" onBack={() => router.push(LINK_HOME)}>
       <LoginForm />
     </AuthContainer>
   )
