+++ apps/web/module/Authentication/components/LoginForm.tsx
@@ -11,7 +11,7 @@ import useLogin from "@/module/Authentication/hooks/useLogin2"
 import { useRouter, useSearchParams } from "next/navigation"
 import { Button } from "@/common/components/ui/Button"
 import Link from "next/link"
-import { LINK_CREATE_ACCOUNT } from "@/common/constants/links"
+import { LINK_CREATE_ACCOUNT } from "@/common/constants"
 import { CONTINUE } from "@/common/constants"
 import Image from "next/image"
 import { Input } from "@/common/components/ui/Input"
